+++ src/EmailHelper.php
@@ -45,8 +45,11 @@ class EmailHelper
         }
 
         $text = self::collapseReplies($text, $enabled);
-        $text = nl2br($text);
-        $text = Link_Filter::activateLinks($text);
+
+        if (!Link_Filter::markdownEnabled()) {
+            $text = nl2br($text);
+            $text = Link_Filter::activateLinks($text);
+        }
 
         return $text;
     }
@@ -64,7 +67,11 @@ class EmailHelper
         }
 
         $wrapText = function ($text) {
-            $text = Misc::highlightQuotedReply($text);
+            if (Link_Filter::markdownEnabled()) {
+                $text = Link_Filter::markdownFormat($text);
+            } else {
+                $text = Misc::highlightQuotedReply($text);
+            }
 
             return
                 '<div><span class="toggle-trimmed-email"><a href="#">…</a></span>'
@@ -85,6 +92,10 @@ class EmailHelper
             function (Fragment $fragment) use ($wrapText) {
                 if ($fragment->isQuoted() || $fragment->isSignature()) {
                     return $wrapText($fragment);
+                }
+
+                if (Link_Filter::markdownEnabled()) {
+                    return Link_Filter::markdownFormat((string)$fragment);
                 }
 
                 return htmlspecialchars($fragment);